Se cambiaron los tipos de los id, se reestablecieron a long pq todo el modelado se hizo en base al id con tipo de dato long

+++ MailboxAPI/Models/DataManager/FacturaManager.cs
@@ -24,7 +24,7 @@ namespace MailboxAPI.Models.DataManager
 
         public long Delete(long id)
         {
-            int facturaID = 0;
+            long facturaID = 0;
             var factura = ctx.Facturas.FirstOrDefault(b => b.FacturaId == id);
             if (factura != null)
             {
